refactor(helpers): rename and move the package manifest helper functions to a single file
