+++ src/cli/actions/generate.action.ts
@@ -5,7 +5,7 @@ import ora from 'ora';
 import chalk from 'chalk';
 import {EXTENSION_LANGUAGES, type NestedExtensionType} from '@directus/extensions';
 import {DirectusRunner, PackageManagerRunner} from '@lib/runners';
-import {detectPackageManager, updateBuildPackageManifestEntries} from '@lib/helpers';
+import {detectPackageManager, updateBundleExtensionEntriesInBuildPackageManifest} from '@lib/helpers';
 
 export default class GenerateAction extends AbstractAction {
 	async handle(inputs?: CommandInput[], options?: CommandInput[]): Promise<void> {
@@ -38,7 +38,7 @@ export default class GenerateAction extends AbstractAction {
 
 			await new DirectusRunner().setAddCommand().setExecaOptions({preferLocal: true}).setPromptAnswers(promptAnswers).build().run();
 
-			await updateBuildPackageManifestEntries(projectPath);
+			await updateBundleExtensionEntriesInBuildPackageManifest(projectPath);
 
 			if (install) {
 				spinner.succeed().start(chalk.bold('Installing dependencies'));
