+++ src/cli/actions/new.action.ts
@@ -7,7 +7,7 @@ import {EXTENSION_TYPES} from '@directus/extensions';
 import {DirectusRunner, GitRunner, PackageManagerRunner} from '@lib/runners';
 import chalk from 'chalk';
 import fse from 'fs-extra';
-import {createBuildPackageManifest, createDirectories, getDoneMessage, getTemplatePath, updatePackageManifest} from '@lib/helpers';
+import {createBuildPackageManifest, createDirectories, getDoneMessage, getTemplatePath, updateBuildDirPathInProjectPackageManifest} from '@lib/helpers';
 import {BUILD_DIRNAME, DIRECTUS_EXTENSION_FOLDER_PREFIX} from '@common/constants.js';
 
 export default class NewAction extends AbstractAction {
@@ -46,7 +46,7 @@ export default class NewAction extends AbstractAction {
 			await createBuildPackageManifest(targetPath,`${targetPath}/${BUILD_DIRNAME}/${DIRECTUS_EXTENSION_FOLDER_PREFIX}${targetDir}`);
 
 			spinner.succeed().start(chalk.bold('updating package manifest'));
-			await updatePackageManifest(targetPath, targetDir);
+			await updateBuildDirPathInProjectPackageManifest(targetPath, targetDir);
 
 			spinner.succeed().start(chalk.bold('copying templates files'));
 			await fse.copy(getTemplatePath(), targetPath, {overwrite: false});
