+++ src/lib/helpers/index.ts
@@ -1,8 +1,6 @@
 export * from './get-cli-version.js';
-export * from './update-package-maniftest.js';
 export * from './get-template-path.js';
 export * from './create-directory.js';
 export * from './get-done-message.js';
 export * from './detect-package-manager.js'
-export * from './create-build-package-manifest.js'
-export * from './update-build-package-manifest-entries.js'
+export * from './package-manifest.js'
